Vertex data uses shorts instead of floats

+++ engine/main.h
@@ -13,8 +13,8 @@
 
 struct VertexData8
 {
-	float x,y,s,t;
-	int atlasId;
+	unsigned short x,y,s,t;
+	unsigned short atlasId;
 };
 
 enum //gamestate
